refactor: make card flexible with its layout
