+++ packages/card/__tests__/card.native.test.js
@@ -4,6 +4,7 @@ import renderer from "react-test-renderer";
 import Card from "../card";
 
 const cardProps = {
+  childRatio: 2.7,
   image: {
     uri:
       "https://www.thetimes.co.uk/imageserver/image/methode%2Fsundaytimes%2Fprod%2Fweb%2Fbin%2F9242e576-4dfc-11e7-a20e-a11097d3353d.jpg?crop=1463%2C975%2C293%2C12"
