+++ packages/card/__tests__/web/card.web.test.js
@@ -5,6 +5,7 @@ import "jest-styled-components";
 import Card from "../../card";
 
 const cardProps = {
+  childRatio: 2.7,
   image: {
     uri:
       "https://www.thetimes.co.uk/imageserver/image/methode%2Fsundaytimes%2Fprod%2Fweb%2Fbin%2F9242e576-4dfc-11e7-a20e-a11097d3353d.jpg?crop=1463%2C975%2C293%2C12"
